restructured authorized package
