+++ app/src/main/java/com/babylon/wallet/android/presentation/dapp/authorized/InitialAuthorizedLoginRoute.kt
@@ -1,4 +1,4 @@
-package com.babylon.wallet.android.presentation.dapp
+package com.babylon.wallet.android.presentation.dapp.authorized
 
 import com.babylon.wallet.android.domain.model.messages.RequiredPersonaFields
 import com.radixdlt.sargon.AccountAddress
